optimize authorization module boundaries

+++ mango/mango-platform/mango-authorization/mango-authorization-support/src/test/java/io/mango/authorization/security/core/impl/DefaultPermissionServiceImplTest.java
@@ -39,7 +39,7 @@ class DefaultPermissionServiceImplTest {
     @Test
     @DisplayName("listUserPermissions with null userId should return empty list")
     void listUserPermissions_nullUserId_returnsEmptyList() {
-        List<String> result = service.listUserPermissions(null);
+        List<String> result = service.listUserPermissions((Long) null);
         assertTrue(result.isEmpty());
     }
